Move java sources to org.wso2.carbon.identity.application.authentication.endpoint.util from authentication endpoint.
