+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.endpoint.util/src/main/java/org/wso2/carbon/identity/application/authentication/endpoint/util/client/exception/ServiceClientException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.application.authentication.endpoint.client.exception;
+package org.wso2.carbon.identity.application.authentication.endpoint.util.client.exception;
 
 /**
  * Exception class handling exceptions at service invocations.
@@ -24,14 +24,17 @@ package org.wso2.carbon.identity.application.authentication.endpoint.client.exce
 public class ServiceClientException extends Exception {
 
     public ServiceClientException(String message) {
+
         super(message);
     }
 
     public ServiceClientException(String message, Throwable cause) {
+
         super(message, cause);
     }
 
     public ServiceClientException(Throwable cause) {
+
         super(cause);
     }
 }
